OpenWeatherMap Connector

+++ kafka-1a/kafka-connector-1/src/main/java/com/simon/kafka/validators/LatLonValidator.java
@@ -7,7 +7,7 @@ public class LatLonValidator implements ConfigDef.Validator {
 
     @Override
     public void ensureValid(String name, Object value) {
-        Integer LatLon = (Integer) value;
+        Double LatLon = (Double) value;
         if(name=="longitude"){
             if (!(-180 <= LatLon && LatLon <= 180)){
                 throw new ConfigException(name, value, "Invalid coordinate.");
